--- Examples/SlidersExamples macOS/AppDelegate.swift
@@ -1,32 +1,0 @@
-import Cocoa
-import SwiftUI
-
-@NSApplicationMain
-class AppDelegate: NSObject, NSApplicationDelegate {
-
-    var window: NSWindow!
-
-
-    func applicationDidFinishLaunching(_ aNotification: Notification) {
-        // Insert code here to initialize your application
-        window = NSWindow(
-            contentRect: NSRect(x: 0, y: 0, width: 480, height: 300),
-            styleMask: [.titled, .closable, .miniaturizable, .resizable, .fullSizeContentView],
-            backing: .buffered, defer: false)
-        window.center()
-        window.minSize = CGSize(width: 480, height: 300)
-        window.setFrameAutosaveName("Main Window")
-
-        let content = ContentView().environmentObject(Model())
-        window.contentView = NSHostingView(rootView: content)
-
-        window.makeKeyAndOrderFront(nil)
-    }
-
-    func applicationWillTerminate(_ aNotification: Notification) {
-        // Insert code here to tear down your application
-    }
-
-
-}
-
